refactor(compiler): split output codegen by sink
`output.rs` grouped its functions by kind, so a section headed "File row
formatter" held three stdout helpers, and the two preambles sat together
under a heading that named neither sink. The call graph splits cleanly
the other way: `data_field_accessors` and `stdout_accessor` are reachable
only from the stdout row writer, `gen_row_bytes` and `gen_value_bytes`
serve only the two file paths.

It is now `output/`, one module per sink behind a routing `mod.rs`:
`merge` selects the sink and sequences the blocks, while `file` and
`stdout` each own a preamble, a row format and their own flushing. The
six functions `merge` calls across the boundary are `pub(super)`; every
other helper is private to its sink.

`gen_merge_section` built its blocks twice, once in each branch. It now
makes a single pass over the IDBs in declaration order, collecting each
relation's drain and count, so only the sequencing differs by sink.

Moving the code brought it under the comment rules it had been exempt
from: twelve lines carrying non-ASCII characters and six over-long
comment lines (rules 7 and 11), a section title that restated the file's
own name, and three facts each stated in two places.
